+++ SkyModel/Sky/ModSMFromFITS.py
@@ -32,10 +32,10 @@ def ReadSMFromFITS(infile):
     Cf=c.data
 
 
-    Cat=np.zeros((ns,),dtype=[('Name','|S200'),('ra',np.float),('dec',np.float),('Sref',np.float),('I',np.float),('Q',np.float),\
-                              ('U',np.float),('V',np.float),('RefFreq',np.float),('alpha',np.float),('ESref',np.float),\
-                              ('Ealpha',np.float),('kill',np.int),('Cluster',np.int),('Type',np.int),('Gmin',np.float),\
-                              ('Gmaj',np.float),('Gangle',np.float),("Select",np.int),('l',np.float),('m',np.float),("Exclude",bool)])
+    Cat=np.zeros((ns,),dtype=[('Name','|S200'),('ra',float),('dec',float),('Sref',float),('I',float),('Q',float),\
+                              ('U',float),('V',float),('RefFreq',float),('alpha',float),('ESref',float),\
+                              ('Ealpha',float),('kill',int),('Cluster',int),('Type',int),('Gmin',float),\
+                              ('Gmaj',float),('Gangle',float),("Select",int),('l',float),('m',float),("Exclude",bool)])
     Cat=Cat.view(np.recarray)
     Cat.RefFreq=1.
     Cat.ra[0:ns]=Cf.RA
